chore: fix detekt issues
