+++ app/src/main/java/com/picpay/desafio/android/data/local/LocalConfigs.kt
