+++ core-network/src/main/java/com/picpay/desafio/android/coreNetwork/retrofit/client/NetworkManager.kt
@@ -1,0 +1,28 @@
+package com.picpay.desafio.android.coreNetwork.retrofit.client
+
+import android.content.Context
+import android.net.ConnectivityManager
+import android.net.NetworkCapabilities
+import android.os.Build
+
+class NetworkManager(private val context: Context) {
+
+    fun isNetworkAvailable(): Boolean {
+        val connectivityManager =
+            context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+            val capabilities =
+                connectivityManager.getNetworkCapabilities(connectivityManager.activeNetwork)
+            if (capabilities != null) {
+                return capabilities.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) ||
+                        capabilities.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) ||
+                        capabilities.hasTransport(NetworkCapabilities.TRANSPORT_ETHERNET)
+
+            }
+        }
+        val activeNetworkInfo = connectivityManager.activeNetworkInfo
+        return (activeNetworkInfo != null && activeNetworkInfo.isConnected)
+    }
+
+}
